+++ changelog.md
@@ -19,8 +19,10 @@ This file is updated nearly every commit and copied to GH release changelog.
 * (internal API) py2noug can now properly convert lists and tuples containing python values
 * (internal API) py2noug can now properly convert dicts, under a list of [key, value] lists
 * (internal API) add an alias `is_noug_num` to `is_n_num` function
+* (internal API) add an alias `is_tok_type` to `does_tok_type_exist` function
 * `«` and `»` are now valid string delimiters
 * Fix some bugs, such as #18
+* Moved config files to the right directory depending on the OS (such as `~/.config`). More infos in the docstring of the src.conffiles file.
 
 ### Calculator
 * Now, the stack is kept at the end of a command. You can make computations using multiple commands, like in Unix’ `dc`.